requirements.txt
# Core dependencies
python-dotenv~=1.0.1

# Deep learning

# Reinforcement learning

# API and data processing

# Model interpretation

# Utilities
uvicorn~=0.32.1
fastapi~=0.115.5
setuptools~=75.8.2
pydantic~=2.9.2
requests~=2.30.0